Create regression_model_for_predicting_the_car_fuel_efficiency.md
This project is about the creation of a model for helping users to predict car prices. The dataset was obtained from this kaggle competition. It is a part of MLZOOMCAMP 2025

Project plan:

Prepare data and Exploratory data analysis (EDA)
Use linear regression for predicting price
Understanding the internals of linear regression
Evaluating the model with RMSE
Feature engineering
Regularization
Using the model